requirements.txt
discord.py
asyncio
torch
openai
youtube-dl